s7e2/refactoring and componentization
